+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # The LabKey Remote API Library for Java - Change Log
 
-## version 5.2.0-SNAPSHOT
+## version 5.3.0-SNAPSHOT
 *Released*: TBD
+
+## version 5.2.0
+*Released*: 3 May 2023
+* Add `RenameContainerCommand` and `RenameContainerResponse`
 
 ## version 5.1.0
 *Released*: 3 March 2023